check if eclipse is already installed

+++ eclipse/install.sh
@@ -8,18 +8,20 @@ FILE=".tmp/eclipse.tar.gz"
 
 info "Installing $NAME"
 
-download_source "$URL" "$FILE"
-extract_tar "$FILE" "/opt/"
-sudo sh -c 'echo "[Desktop Entry]
-Name=Eclipse
-Type=Application
-Exec=/opt/eclipse/eclipse
-Terminal=false
-Icon=/opt/eclipse/icon.xpm
-Comment=Integrated Development Environment
-NoDisplay=false
-Categories=Development;IDE;
-Name[en]=Eclipse" >> "/usr/share/applications/eclipse.desktop"'
-rm "$FILE"
+if test ! "[ -f /usr/share/applications/eclipse.desktop ]"; then
+  download_source "$URL" "$FILE"
+  extract_tar "$FILE" "/opt/"
+  sudo sh -c 'echo "[Desktop Entry]
+  Name=Eclipse
+  Type=Application
+  Exec=/opt/eclipse/eclipse
+  Terminal=false
+  Icon=/opt/eclipse/icon.xpm
+  Comment=Integrated Development Environment
+  NoDisplay=false
+  Categories=Development;IDE;
+  Name[en]=Eclipse" >> "/usr/share/applications/eclipse.desktop"'
+  rm "$FILE"
+fi
 
 success "$NAME installed"
